Fix(build): Set static version and bump Java to 17
Set a static `versionCode` and `versionName` by removing the dynamic versioning based on Git commit count.

Upgraded the project's Java and Kotlin compatibility to version 17. This includes updating `compileOptions`, setting the `jvmToolchain` to 17, and removing outdated `kotlinOptions`.

Additionally, the `isMinifyEnabled` flag for release builds is now set to `true`, and the `org.gradle.toolchains.foojay-resolver-convention` plugin was added to `settings.gradle.kts` to standardize JDK management.